ai-service/requirements.txt
fastapi
uvicorn
requests
faker
python-dotenv
pydantic
